refactor: deduplicate against pre-existing repo code
From REDUNDANT.md section 3:

- run_checked() now builds on imagecraft.subprocesses.run instead of
  re-implementing subprocess plumbing.
- build_prime_chroot() moved from bootloader/chrootenv.py into the
  existing imagecraft/pack/chroot.py module.
- Partition discovery is now on the volume models themselves:
  BaseVolume gains root_partition/esp_partition/boot_partition and
  has_bios_boot_partition properties (via a shared _VolumePartitions
  mixin), replacing the find_*_structure_item helpers in installer.py
  and the repeated scans in bios.py.
- GRUB module staging deduplicated into chrootenv.stage_grub_modules,
  used by both the BIOS staging step and the EFI fallback build.
