+++ imagecraft/pack/chroot.py
@@ -192,3 +192,52 @@ class Chroot:
             raise errors.ChrootExecutionError(err)
 
         return res
+
+
+def build_prime_chroot(
+    root_dir: Path,
+    *,
+    boot_dir: Path | None = None,
+    extra_mounts: list[Mount] | None = None,
+) -> Chroot:
+    """Build a chroot rooted at a partition's prime directory.
+
+    Only the device files GRUB tooling needs are bind-mounted (rather than
+    overmounting ``/dev``, which would hide the bind targets).
+
+    :param root_dir: Prime directory of the root filesystem partition.
+    :param boot_dir: Prime directory of a dedicated ``/boot`` partition,
+        bound at ``/boot`` in the chroot. Defaults to the root partition's
+        own ``/boot`` when not given.
+    :param extra_mounts: Additional mounts to set up inside the chroot
+        (e.g. tool shims bind-mounted over the guest's binaries).
+    """
+    for mountpoint in ("proc", "sys", "dev", "tmp"):
+        (root_dir / mountpoint).mkdir(parents=True, exist_ok=True)
+
+    mounts = [
+        Mount(fstype="proc", src="proc-build", relative_mountpoint="/proc"),
+        Mount(fstype="sysfs", src="sysfs-build", relative_mountpoint="/sys"),
+    ]
+    for device in ("null", "zero", "urandom"):
+        (root_dir / "dev" / device).touch(exist_ok=True)
+        mounts.append(
+            Mount(
+                fstype=None,
+                src=f"/dev/{device}",
+                relative_mountpoint=f"/dev/{device}",
+                options=["--bind"],
+            )
+        )
+    if boot_dir is not None:
+        (root_dir / "boot").mkdir(exist_ok=True)
+        mounts.append(
+            Mount(
+                fstype=None,
+                src=str(boot_dir.resolve()),
+                relative_mountpoint="/boot",
+                options=["--bind"],
+            )
+        )
+    mounts.extend(extra_mounts or [])
+    return Chroot(path=root_dir, mounts=mounts)
